refactor mod imports for vault
