+++ contracts/vault/src/lib.rs
@@ -1,24 +1,18 @@
-mod types;
-pub use types::*;
-
-mod contract;
-pub use contract::Vault;
-
 mod claim_unstaked;
+mod contract;
 mod delegate;
+mod ext;
 mod ft_receiver;
 mod internal;
+mod request_liquidity;
 mod transfer_ownership;
+mod types;
 mod undelegate;
 mod view;
 mod withdraw_balance;
-mod request_liquidity;
 
 #[cfg(test)]
 mod unit;
 
 #[macro_use]
 mod macros;
-
-mod ext;
-pub use ext::*;
