+++ contracts/vault/src/undelegate.rs
@@ -1,12 +1,17 @@
+#![allow(dead_code)]
+
+use crate::contract::Vault;
 use crate::contract::VaultExt;
-use crate::ext_self;
+use crate::ext::ext_self;
+use crate::ext::{
+    METHOD_GET_ACCOUNT_STAKED_BALANCE, METHOD_GET_ACCOUNT_UNSTAKED_BALANCE, METHOD_UNSTAKE,
+    METHOD_WITHDRAW_ALL,
+};
 use crate::log_event;
-use crate::types::*;
-use crate::Vault;
-use crate::METHOD_GET_ACCOUNT_STAKED_BALANCE;
-use crate::METHOD_GET_ACCOUNT_UNSTAKED_BALANCE;
-use crate::METHOD_UNSTAKE;
-use crate::METHOD_WITHDRAW_ALL;
+use crate::types::{
+    StorageKey, UnstakeEntry, GAS_FOR_CALLBACK, GAS_FOR_UNSTAKE, GAS_FOR_VIEW_CALL,
+    GAS_FOR_WITHDRAW_ALL,
+};
 use near_sdk::collections::Vector;
 use near_sdk::json_types::U128;
 use near_sdk::{assert_one_yocto, env, near_bindgen, AccountId, NearToken, Promise};
