+++ contracts/vault/src/unit/mod.rs
@@ -7,3 +7,4 @@ mod test_utils;
 mod transfer_ownership;
 mod undelegate;
 mod withdraw_balance;
+mod request_liquidity_tests;
