+++ contracts/vault/src/unit/reconcile_unstake_entries.rs
@@ -1,6 +1,10 @@
 #[path = "test_utils.rs"]
 mod test_utils;
-use crate::{StorageKey, UnstakeEntry, Vault};
+
+use crate::{
+    contract::Vault,
+    types::{StorageKey, UnstakeEntry},
+};
 use near_sdk::{collections::Vector, env, testing_env, AccountId, NearToken};
 use test_utils::{get_context, owner};
