+++ contracts/vault/src/unit/request_liquidity_tests.rs
@@ -1,0 +1,2 @@
+#[path = "test_utils.rs"]
+mod test_utils;
